Accept ready-made layered images as a second order path
Customers who already built a layered image elsewhere can now order it
directly. The studio page carries a mode switch between building an image
with our pipeline and uploading a finished one.

Verification (posterizer/ready.py)
- An upload is accepted only if it is genuinely made of solid colours.
  Rather than counting exact colours — a four-colour PNG saved as JPEG has
  thousands — pixels are clustered greedily around their most frequent
  colours, tiny clusters are folded into the nearest real layer, and only
  then is the layer count judged. Judging raw clusters would reject clean
  artwork purely for having anti-aliased edges.
- Gradients, photographs and near-single-colour images are refused with a
  specific Persian reason rather than being silently altered.
- The palette is read from the image itself; no colour profile is involved.
  The stored copy is snapped onto that palette so compression noise cannot
  become stray terraces in the 3D model.
- Thresholds (max colours, minimum coverage, minimum layer share) are all
  admin-configurable, and the whole path can be switched off.

Shared behaviour
- Sizing, pricing, the confirmation dialog, the three-order quota and STL
  export are identical on both paths. Orders record their source, shown in
  the admin and tagged on the orders page. Switching back to the studio
  clears the verified image so the paths cannot mix.

AI helper
- A card beside the upload area tells the customer how to produce a suitable
  image: send their photo to a named model with a ready-written prompt, then
  upload the result. Model name, link and prompt text are editable in the
  admin, with a one-click copy button.

Fixes
- Colour matching overflowed int16. A squared channel difference reaches
  65025, which wraps negative, so argmin could return a completely wrong
  palette entry: a pixel exactly equal to layer 0 was mapped to layer 3.
  This silently corrupted STL height assignment. Both the exporter and the
  new detector now use int32, with a regression test.
- format_html() with no interpolation arguments raises TypeError on modern
  Django; it crashed the order list for any order with an empty palette.

Docs updated throughout. 72 tests pass.

+++ posterizer/stl.py
@@ -44,8 +44,10 @@ def layer_index_map(image: Image.Image, colors, max_resolution: int) -> np.ndarr
         new_size = (max(1, round(width * scale)), max(1, round(height * scale)))
         image = image.resize(new_size, Image.NEAREST)
 
-    pixels = np.asarray(image.convert("RGB"), dtype=np.int16)
-    palette = np.array([hex_to_rgb(color) for color in colors], dtype=np.int16)
+    # int32, not int16: a squared channel difference reaches 255**2 = 65025,
+    # which overflows int16 and makes argmin pick the wrong colour entirely.
+    pixels = np.asarray(image.convert("RGB"), dtype=np.int32)
+    palette = np.array([hex_to_rgb(color) for color in colors], dtype=np.int32)
 
     # Nearest palette entry per pixel, so a stray colour still lands on a layer.
     distances = np.sum((pixels[:, :, None, :] - palette[None, None, :, :]) ** 2, axis=3)
